requirements.txt
certifi
paho-mqtt
